Redefined code and fixed bugs related db

+++ Commands/management.py
@@ -45,11 +45,10 @@ class Management(commands.Cog):
     @commands.has_permissions(administrator=True)
     @commands.command(name="stats disable", aliases=("sd",), help="Removes stats from the server")
     async def stats_disable(self, ctx):
-        global find, status
+        global status
         global category_stats_id
         for x in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "category stats id": 1}):
             category_stats_id = get_response(x)
-            find = x
 
         for y in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "stats status": 1}):
             status = get_response(y)
@@ -70,8 +69,9 @@ class Management(commands.Cog):
 
             disable_stats = {"$set": {"stats status": False, "members channel id": 0,
                                       "bots channel id": 0, "category stats id": 0}}
-            if find is not None:
-                server_collection.update_one(find, disable_stats)
+            find_server = {"guild id": ctx.guild.id}
+            if find_server is not None:
+                server_collection.update_one(find_server, disable_stats)
             await ctx.reply(f"Successfully disabled stats")
         else:
             await ctx.reply(f"Stats are already disabled")
@@ -79,15 +79,24 @@ class Management(commands.Cog):
     @commands.has_permissions(administrator=True)
     @commands.command(name="quotes enable", aliases=("qe",), help="Adds daily quotes to the server")
     async def quotes_enable(self, ctx):
-        global status
+        global status, category_info_id
         for x in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "daily quotes status": 1}):
             status = get_response(x)
             print(status)
+
         if status == "False":
             overwrites_text_channel = {
                 ctx.guild.default_role: discord.PermissionOverwrite(send_messages=False)
             }
-            category_info = await ctx.guild.create_category(name="↢↢↢↢↢『DailyDose』↣↣↣↣↣", position=0)
+            try:
+                for y in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "category info id": 1}):
+                    category_info_id = get_response(y)
+                    print(status)
+                category_info = discord.Guild.get_channel(ctx.guild, int(category_info_id))
+            except Exception as e:
+                print(f"Error in fetching category: {e} ")
+            if category_info is None:
+                category_info = await ctx.guild.create_category(name="↢↢↢↢↢『DailyDose』↣↣↣↣↣", position=0)
             quote_channel = await ctx.guild.create_text_channel(name=f"『📚』: Daily-Quotes",
                                                           overwrites=overwrites_text_channel,
                                                           category=category_info, news=True)
@@ -96,17 +105,16 @@ class Management(commands.Cog):
                                      "category info id": category_info.id}}
             if find_server is not None:
                 server_collection.update_one(find_server, enable_stats)
-            await ctx.reply(f"Successfully enabled daily quotes")
+                await ctx.reply(f"Successfully enabled daily quotes")
         else:
             await ctx.reply(f"Daily quotes are already enabled")
 
     @commands.has_permissions(administrator=True)
     @commands.command(name="quotes disable", aliases=("qd",), help="Removes daily quotes from the server")
     async def quotes_disable(self, ctx):
-        global find, status, history_channel_id, quote_channel_id, category_info_id
+        global status, history_channel_id, quote_channel_id, category_info_id
         for x in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "category info id": 1}):
             category_info_id = get_response(x)
-            find = x
 
         for y in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "daily quotes status": 1}):
             status = get_response(y)
@@ -129,13 +137,14 @@ class Management(commands.Cog):
         if status == "True":
             try:
                 info_category = discord.Guild.get_channel(ctx.guild, int(category_info_id))
-                if quote_channel_id is not None:
+                if quote_channel_id != "0" or not None:
                     quote_channel = discord.Guild.get_channel(ctx.guild, int(quote_channel_id))
-                if history_channel_id is not None:
+                if history_channel_id != "0" or not None:
                     history_channel = discord.Guild.get_channel(ctx.guild, int(history_channel_id))
-                for channel in info_category.text_channels:
-                    if channel == quote_channel:
-                        await channel.delete()
+                if info_category != "0" or not None:
+                    for channel in info_category.text_channels:
+                        if channel == quote_channel:
+                            await channel.delete()
 
                 if history_channel is None:
                     await info_category.delete()
@@ -144,34 +153,36 @@ class Management(commands.Cog):
             except Exception as e:
                 await ctx.reply(f"Error Occurred: {e}")
 
-            disable_stats = {"$set": {"stats status": False, "history channel id": 0,
+            disable_stats = {"$set": {"daily quotes status": False, "quotes channel id": 0,
                                       "category info id": int(category_info_id)}}
-            if find is not None:
-                server_collection.update_one(find, disable_stats)
-            await ctx.reply(f"Successfully disabled daily history")
+            find_server = {"guild id": ctx.guild.id}
+            if find_server is not None:
+                server_collection.update_one(find_server, disable_stats)
+                await ctx.reply(f"Successfully disabled daily history")
         else:
             await ctx.reply(f"Daily history is already disabled")
 
     @commands.has_permissions(administrator=True)
     @commands.command(name="history enable", aliases=("he",), help="Adds daily history to the server")
     async def history_enable(self, ctx):
-        global status
+        global status, category_info_id, category_info
         for x in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "daily history status": 1}):
             status = get_response(x)
             print(status)
         if status == "False":
             overwrites_text_channel = {
-                ctx.default_role: discord.PermissionOverwrite(send_messages=False)
+                ctx.guild.default_role: discord.PermissionOverwrite(send_messages=False)
             }
             try:
                 for y in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "category info id": 1}):
-                    status = get_response(y)
+                    category_info_id = get_response(y)
                     print(status)
                 category_info = discord.Guild.get_channel(ctx.guild, int(category_info_id))
             except Exception as e:
                 print(f"Error in fetching category: {e} ")
-                category_info = await ctx.create_category(name="↢↢↢↢↢『DailyDose』↣↣↣↣↣", position=0)
-            history_channel = await ctx.create_text_channel(name=f"『🤔』: Daily-History",
+            if category_info is None:
+                category_info = await ctx.guild.create_category(name="↢↢↢↢↢『DailyDose』↣↣↣↣↣", position=0)
+            history_channel = await ctx.guild.create_text_channel(name=f"『🤔』: Daily-History",
                                                             overwrites=overwrites_text_channel,
                                                             category=category_info, news=True)
             find_server = {"guild id": ctx.guild.id}
@@ -179,7 +190,7 @@ class Management(commands.Cog):
                                      "category info id": category_info.id}}
             if find_server is not None:
                 server_collection.update_one(find_server, enable_stats)
-            await ctx.reply(f"Successfully enabled daily history")
+                await ctx.reply(f"Successfully enabled daily history")
         else:
             await ctx.reply(f"Daily history is already enabled")
 
@@ -187,38 +198,37 @@ class Management(commands.Cog):
     @commands.command(name="history disable", aliases=("hd",), help="Removes daily history from the server")
     async def history_disable(self, ctx):
         global find, status, history_channel_id, quote_channel_id, category_info_id
-        for x in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "category info id": 1}):
-            category_info_id = get_response(x)
-            find = x
-
         for y in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "daily history status": 1}):
             status = get_response(y)
 
-        try:
-            for history_channel in server_collection.find({"guild id": ctx.guild.id},
-                                                          {"_id": 0, "history channel id": 1}):
-                history_channel_id = get_response(history_channel)
-        except Exception as e:
-            print(f"Error occurred: {e}")
-            history_channel_id = None
-
-        try:
-            for quote_channel in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "quote channel id": 1}):
-                quote_channel_id = get_response(quote_channel)
-        except Exception as e:
-            print(f"Error occurred: {e}")
-            quote_channel_id = None
-
         if status == "True":
+            for x in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "category info id": 1}):
+                category_info_id = get_response(x)
+                find = x
+            try:
+                for history_channel in server_collection.find({"guild id": ctx.guild.id},
+                                                              {"_id": 0, "history channel id": 1}):
+                    history_channel_id = get_response(history_channel)
+            except Exception as e:
+                print(f"Error occurred: {e}")
+                history_channel_id = None
+
+            try:
+                for quote_channel in server_collection.find({"guild id": ctx.guild.id}, {"_id": 0, "quote channel id": 1}):
+                    quote_channel_id = get_response(quote_channel)
+            except Exception as e:
+                print(f"Error occurred: {e}")
+                quote_channel_id = None
             try:
                 info_category = discord.Guild.get_channel(ctx.guild, int(category_info_id))
                 if quote_channel_id is not None:
                     quote_channel = discord.Guild.get_channel(ctx.guild, int(quote_channel_id))
                 if history_channel_id is not None:
                     history_channel = discord.Guild.get_channel(ctx.guild, int(history_channel_id))
-                for channel in info_category.text_channels:
-                    if channel == history_channel:
-                        await channel.delete()
+                if info_category is not None:
+                    for channel in info_category.text_channels:
+                        if channel == history_channel:
+                            await channel.delete()
 
                 if quote_channel is None:
                     await info_category.delete()
@@ -227,11 +237,12 @@ class Management(commands.Cog):
             except Exception as e:
                 await ctx.reply(f"Error Occurred: {e}")
 
-            disable_stats = {"$set": {"stats status": False, "history channel id": 0,
+            disable_stats = {"$set": {"daily history status": False, "history channel id": 0,
                                       "category info id": int(category_info_id)}}
-            if find is not None:
-                server_collection.update_one(find, disable_stats)
-            await ctx.reply(f"Successfully disabled daily history")
+            find_server = {"guild id": ctx.guild.id}
+            if find_server is not None:
+                server_collection.update_one(find_server, disable_stats)
+                await ctx.reply(f"Successfully disabled daily history")
         else:
             await ctx.reply(f"Daily history is already disabled")
